table: let major compaction wait on background deletions
In several cases in distributed testing (dtest) we trigger compaction using nodetool compact assuming that when it is done, it is indeed really done.
However, the way compaction is currently implemented in scylla, it may leave behind some background tasks to delete the old sstables that were compacted.

This commit changes major compaction (triggered via the ss::force_keyspace_compaction api)
so it would wait on the background deletes and will return only when they finish.

Fixes #4909

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>