sstables: create_links: support for move
When moving a sstable between directories, we would like to
atomically rename the TemporaryTOC component that is created while
the components are being linked back to the source directory,
as the latter is about to be used for removing the source sstable
components.  This way, if we crash and restart at any point during
the process, we will clean up either the source or destination sstable.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>