sstables: Add support for the "md" format
Scylla does not suffer from the issue detailed in CASSANDRA-14861, which
lead to the creation of the "md" format. Currenly, Scylla can,
therefore, consume the "md" format files, except that users need to
rename the files to have a "mc" suffix. Let's fix that up by detecting
"md" suffix, but treating it like "mc" internally.

Fixes: #4442

[sstables: fix md format support]
[test: fixed unit tests for md format]
[test: added md sstables for unit tests]
[use >= instead of sstable_version_types::is_at_least]
Signed-off-by: Benny Halevy <bhalevy@scylladb.com>