table: filter_sstable_for_reader: adjust to md-format
With the md-format, with Scylla's implementation that reflects also
partition tombstones in min/max column names, the refer to
clustering key prefixes of clustered rows, range tombstone bounds,
and implicitly (as infinite bounds) - to partition tombstones.

As such, min and max coloumn names may be of different lengths
due to range tombstones, so the current matching algorithm must
be changed to take this into account.

To determine if a slice range overlaps the min/max range
we need to use clustering_key_prefix::prefix_equal_tri_compare
so that a partial min/max prefix will match a longer slice
key as equal. Otherwise, additional components in the
slice range may appear to be smaller than min_coloum_names
(with descending order) or larger than max_column_names
(with ascending order), while we need to compare only
the shorter common prefix.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>