sstables: metadata_collector: construct with schema
Pass the schema to the metadata_collector constructor
and check if it has a clustering key only once, rather
than each time update_min_max_components is being called.

Note that the long term plan is to move metadata_collector
to the sstable writer but this requires a bigger change to
get rid of the dependencies on it in the legacy writer code
in class sstable methods.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>