sstables: keep whole clustering_key_prefix as min/max_column_names
Currently we compare each min/max component independently.
This may lead to suboptimal, inclusize clustering ranges
that do not indicate any actual key we encounters.

For example: ['a', 2], ['b', 1] will lead to min=['a', 1], max=['b', 2]
instead of the keys themselves.

This change keeps the min or max keys as a whole.

Also, since we don't track partial key-prefix as possible with
range tombstones, we do not allow unset suffixes starting at this commit.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>