sstables: random_access_reader: unify input stream close code
Let seek() and close() close close_if_needed()
to close the input stream.

The latter captures the input stream unique ptr
so it will remain valid throughout close.
This was missing from close().

Also, a future patch will call it possibly
with disengaged _in so it will close `in` only
if it is engaged.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>