sstables: store ancestors in scylla metadata component
The ancestors metadata was deprecated from compaction
metadata by Cassandra in the mc sstable format.

Restore this information in the scylla metadata component
as it is useful to determine if sstables were created by
compaction or not, e.g. they were flushed from the memtable
or uploaded by nodetool refresh.

Fixes #6398

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>