test: mutation_reader_test: sst_factory: use set_sstable_level
sst_factory currently uses the sstable's metadata collector
sstable_level() to set the sstable's level. However, this method
is intended to be used when writing the sstable, but sst_factory
is returning a sstable for read using make_sstable(), so it is
more correct to set the sstable level using sstable::set_sstable_level.

Also, the metadata collector is moving out of the sstable class
to the sstable writer and it won't be available here.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>