sstables: mc/writer: disable_min_max_components for partition tombstone
The partition tombstone represents an implicit clustering range
that is unbound on both sides, so reflect than in min/max
coloumn names metadata.

Otherwise, when using the sstable for filtering, we have no
way of distinguishing range tombstones from partition tombstones
given the sstable metadata and we would need to include any sstable
with tombstones, even if those are range tombstone, for which
we can do a better filtering job, using the sstable min/max
coloumn names metadata.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>