sstables: make compaction_strategy_impl::get_backlog_tracker non virtual
This was duplicated in every "real" implementation. The reason seems
to be to avoid a new when creating a null_compaction_strategy, which
seems like the wrong thing to optimize for.

This hits problems with _backlog_tracker being constructed to early I
think. That is related to some of the other changes in this
repository.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>