table: create_single_key_sstable_reader: emit partition_start/end for empty filtered results
To prevent https://github.com/scylladb/scylla/issues/3552
if we read a single partition and filter the sstables based on the clustering key
resulting in no sstables with the read partition we want to populate the
cache correctly, hence the reader should emit partition_start/partition_end
for the empty partition if it exists in sstables.

Note that we don't need to emit partition_start/partition_end if
the sstables were filtered by partition key, since we know there
are no sstables containing this partition.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>