storage_service: join_token_ring: replicate_to_all_cores early
To prepare for converting to using shared_token_metadata.
In the new world the updated token_metadata is not visible
to others until explicitly committed, so commit it early
and replicate to all shards.

Currently the updated are immediately visible on shard 0,
but not to other shards until replicate_to_all_cores syncs
them.  I'm not sure if this works by design or by sheer chance.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>