schema_tables: create_table_from_mutations auto detect dropped_time resolution
dropped_columns's type is timestamp_type which is supposed to be in milliseconds since epoch.
We currently wrongly store an api::timestamp_type in it which is in microseconds since epoch.
(See 6c8b5fc09d5e01782acd3ef175113935f2ce9db4)

This change allows building the schema with dropped_columns either in milliseconds
or microseconds.

It does not set the dropped_column.dropped_time in milliseconds yet
until we have a cluster feature to enable it.

Also, add some trace messages to track dropped columns dropped_time.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>