storage_service: make keyspace_changed and update_pending_ranges private
Both are called only internally in the class.

keyspace_changed does not require a storage_service instance
so make it static and use get_storage_service().invoke_on(0
to invoke update_pending_ranges on shard 0.

With that update_pending_ranges() is always called on shard 0
and it doesn't need to use `invoke_on` to run on shard 0.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>