table: snapshot: use sstable_directory::do_for_each_sstable
Tables may have thousands of sstables and a number of component files
for each sstables.  Using parallel_for_each on all sstables (and
parallel_for_each in sstables::create_links for each file)
needlessly overloads the system with unbounded number of continuations.

Use sstable_directory::do_for_each_sstable to limit parallelism to
the configured initial_sstable_loading_concurrency() (that is used
to operate on the sstables directory when loading it).

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>