sstable_directory: support sstables with both TemporaryTOC and TOC
Keep descriptors in a map so it could be searched easily by generation.
and possibly delete the descriptor, if found, in the presence of
a temporary toc component.

We may get into this state when crashing in the middle of
(the revised) create_links.

Add unit test for this case.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>