flat_mutation_reader: introduce abort and close methods
Allow aborting readers, and finally closing them
asynchronously.

abort() may be used to abort the reader
and to exceptions downstream to underlying readers.
Although it's intended to be synchronous it returns
a future so it could submit abort() across shards.

close() is used to asynchronously, and recuresively
close the reader resources before destrying it, allowing
waiting on underlying readers' close.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>