view: close reader when done consuming
Note, we keep the reader shared ptr since it might
be replaced in initialize_reader_at_current_token
while being referenced by reader.consume_in_thread
consumer_adapter.

Currently, it is not an issue, but we want to keep the
reader valid throughout the lifecycle of consumer_adapter
to allow closing it properly.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>