mutation_query: reconcilable_result_builder: mark constructor noexcept
query::result_memory_accounter is already nothrow move constructible,
and the other initializers are trivial.

The motivation for this change is to simplify error handling
in do_mutation_query.
Otherwise, we'll need to handle an exception from
reconcilable_result_builder(*s, slice, std::move(accounter))
and make sure to close the mutation_querier on that path too.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>