reader_concurrency_semaphore: inactive_read: add gentle close
And use it via close_inactive_read that extends
unregister_inactive_read to close the reader
before it's destructed.

Also, void destroy_inactive_read() was added to close
the inactive read in the background, without waiting
for its future<>, so it can be called from the querier
cache that provides a synchronous interface.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>