flat_mutation_reader: detemplatize delegating_reader
row_cache construct a delegating_reader, passing a flat_mutation_reader& to it.
In this case, the delegating_reader does not own the underlying reader
and must not closed it upon close().

Other uses, by row_cache_test use readers derived from delegating_reader.
Convert them to hold the reader as a member and construct a delegating_reader
member rather than be derived from it.

Detemplatize delegating_reader so it is alwasy constructed
with a flat_mutation_reader& that it does not own.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>