flat_mutation_reader: detemplatize delegating_reader
row_cache construct a delegating_reader, passing a flat_mutation_reader& to it.
In this case, the delegating_reader does not own the underlying reader
and must not closed it upon close().

Other uses, by row_cache_test use readers derived from delegating_reader
and they construct it using a moved flat_mutation_reader&&.
In this case the ownership of the underlying reader is moved
to the delegating_reader and the latter is now responsible for
closing it.

To support both use cases, add a flat_mutation_reader_opt
_underlying_holder member that is engaged only when
the _underlying reader is moved to the delegating_reader.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>