mutation_partition: coroutinize data_query and do_mutation_query
In preparation for futurizing the lookup_mutation_querier.

Note that we still close the querier in a finally continuation
to simplify error handling, since coroutines do not support
co_await in a catch clause. Doing it as a defer()ed action
is also nasty since it will have to be done in the background
while ignoring any exceptions.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>