reader_concurrency_semaphore: register_inactive_read: handle errors
Make register_inactive_read exception safe and call
the eviction_notify_handler on any error to register
the inactive_read.

To do that, reverse the order of registration and first
emplace an empty inactive_read in _inactive_reads,
so, if failed, the notify_handler would still be available
to be called.  After that, error handling will be
done using the evict method that also erases the
inactive_read from _inactive_reads.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>