reader_concurrency_semaphore: use a list for inactive_reads
Now that inactive_read_handle holds an iterator to inactive_reads
we can simplify it and use a simple list to hold the inactive_read:s
std::list satisfies the requirement the insert and erase do not
invalidate iterators other than the erased one. This is a must
for keeping iterators in the i_r_h.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>