reader_concurrency_semaphore: do_wait_admission: enqueue_waiter before evicting readers
We admit the read subchronously only if the wait_list is empty
and the semaphore has_available_units to statisfy admission.

Otherwise the reader is enqueued and if it is first in line,
it starts evicting inactive readers.  The reason to do so
is that closing the inactive reads may happen asynchronously
(e.g. with shard_reader::close that may need to be submitted
to a different shard) and we want to claim these resources
in fifo order, hence the waiting reader need to be enqueued
first, before evicting any inactive readers.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>